refactor(connectors): rename warehouse-write-transport → connector-write-transport
Plan anticipated only the dispatch importer; grep found two more (platform/types.ts,
platform/assemble.ts) — all three updated to ConnectorWriteContext.
